+++ test/test14.js
@@ -24,7 +24,7 @@ describe('Test 14', function() {
 		assert.deepEqual([ { a: 1, b: 1, c: 5 }, { a: 1, b: 7, c: 5 } ], res);
 		assert.equal(1,Object.keys(db.tables.test1.indices).length);
 		assert.equal(1,Object.keys(db.tables.test2.indices).length);
-		console.log(db.tables.test1.indices);
+		//console.log(db.tables.test1.indices);
 
 		var res = db.exec('SELECT a,b,test2.c FROM test1 LEFT JOIN test2 ON test1.a = test2.a '+
 			" WHERE test1.a = 2");