+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 0.0.14.4 (10.11.2014)
+
+* Negative numbers 
+* Float numbers
+* Strings with single and double quaters
+
 ### 0.0.14.3 (09.11.2014)
 
 * RIGHT / OUTER / SEMI / ANTI Joins!